DAF Utility functions (#360)
* Utilities moved

* apt
