+++ requirements/docs.txt
@@ -1,4 +1,4 @@
-sphinx==7.0.1
+sphinx==6.2.1
 sphinx-autobuild==2021.3.14
 sphinx-copybutton==0.5.2
 furo==2023.5.20
