+++ src/daf/__init__.py
@@ -2,6 +2,8 @@
 Discord Advertisement Framework
 """
 import _discord as discord
+
+from . import misc
 from .client import *
 from .core import *
 from .dtypes import *
@@ -11,12 +13,7 @@ from .logging import *
 from .web import *
 from .convert import *
 from .remote import *
-from pathlib import Path as __Path
 
-
-from .misc import DOCUMENTATION_MODE
-if DOCUMENTATION_MODE:
-    from .misc import *
 
 
 VERSION = "2.9.0.4"
