+++ src/daf/core.py
@@ -11,10 +11,9 @@ from typeguard import typechecked
 
 from .logging.tracing import TraceLEVELS, trace
 from .logging import _logging as logging, tracing
-
+from .misc import doc, instance_track as it
 from . import guild
 from . import client
-from . import misc
 from . import message
 from . import convert
 from . import remote
@@ -86,7 +85,7 @@ async def http_add_account(account: dict):
 
 @remote.register("/accounts", "DELETE")
 async def http_remove_account(account_id: int):
-    account = misc.get_by_id(account_id)
+    account = it.get_by_id(account_id)
     if account is None:
         raise aiohttp_web.HTTPInternalServerError(reason="Account not present")
 
@@ -161,7 +160,7 @@ async def schema_load_from_file() -> None:
     trace(f"Restored objects from file ({len(GLOBALS.accounts)} accounts).", TraceLEVELS.NORMAL)
 
 
-@misc.doc_category("DAF control reference")
+@doc.doc_category("DAF control reference")
 async def initialize(user_callback: Optional[Union[Callable, Coroutine]] = None,
                      debug: Optional[Union[TraceLEVELS, int, str]] = TraceLEVELS.NORMAL,
                      logger: Optional[logging.LoggerBASE] = None,
@@ -245,7 +244,7 @@ async def initialize(user_callback: Optional[Union[Callable, Coroutine]] = None,
 # Functions
 #######################################################################
 @overload
-@misc.doc_category("Dynamic mod.", True)
+@doc.doc_category("Dynamic mod.", True)
 async def add_object(obj: client.ACCOUNT) -> None:
     """
     Adds an account to the framework.
@@ -266,7 +265,7 @@ async def add_object(obj: client.ACCOUNT) -> None:
 
 
 @overload
-@misc.doc_category("Dynamic mod.", True)
+@doc.doc_category("Dynamic mod.", True)
 async def add_object(obj: Union[guild.USER, guild.GUILD, guild.AutoGUILD],
                      snowflake: client.ACCOUNT = None) -> None:
     """
@@ -295,7 +294,7 @@ async def add_object(obj: Union[guild.USER, guild.GUILD, guild.AutoGUILD],
 
 
 @overload
-@misc.doc_category("Dynamic mod.", True)
+@doc.doc_category("Dynamic mod.", True)
 async def add_object(obj: Union[message.DirectMESSAGE, message.TextMESSAGE, message.VoiceMESSAGE],
                      snowflake: Union[guild.GUILD, guild.USER]) -> None:
     """
@@ -355,7 +354,7 @@ async def add_object(obj, snowflake=None):
 
 
 @typechecked
-@misc.doc_category("Dynamic mod.")
+@doc.doc_category("Dynamic mod.")
 async def remove_object(
     snowflake: Union[guild._BaseGUILD, message.BaseMESSAGE, guild.AutoGUILD, client.ACCOUNT]
 ) -> None:
@@ -399,7 +398,7 @@ async def remove_object(
         GLOBALS.accounts.remove(snowflake)
 
 
-@misc.doc_category("Clients")
+@doc.doc_category("Clients")
 def get_accounts() -> List[client.ACCOUNT]:
     """
     .. versionadded:: v2.4
@@ -413,7 +412,7 @@ def get_accounts() -> List[client.ACCOUNT]:
 
 
 @typechecked
-@misc.doc_category("DAF control reference")
+@doc.doc_category("DAF control reference")
 async def shutdown() -> None:
     """
     Stops and cleans the framework.
@@ -461,7 +460,7 @@ def _shutdown_clean(loop: asyncio.AbstractEventLoop) -> None:
 
 
 @typechecked
-@misc.doc_category("DAF control reference")
+@doc.doc_category("DAF control reference")
 def run(user_callback: Optional[Union[Callable, Coroutine]] = None,
         debug: Optional[Union[TraceLEVELS, int, str, bool]] = TraceLEVELS.NORMAL,
         logger: Optional[logging.LoggerBASE] = None,
