+++ src/daf_gui/connector.py
@@ -5,6 +5,7 @@ clients.
 from typing import List, Optional, Literal
 
 from daf.logging.tracing import TraceLEVELS, trace
+from daf.misc import instance_track as it
 
 from .convert import *
 from .utilities import *
@@ -61,13 +62,13 @@ class AbstractConnectionCLIENT:
         """
         raise NotImplementedError
 
-    async def remove_account(self, account_ref: daf.misc.ObjectReference):
+    async def remove_account(self, account_ref: it.ObjectReference):
         """
         Logs out and removes account from DAF.
 
         Parameters
         ----------------
-        account_ref: daf.misc.ObjectReference
+        account_ref: it.ObjectReference
             The reference to account to remove.
         """
         raise NotImplementedError
@@ -84,7 +85,7 @@ class AbstractConnectionCLIENT:
         """
         raise NotImplementedError
 
-    async def refresh(self, object_ref: daf.misc.ObjectReference) -> object:
+    async def refresh(self, object_ref: it.ObjectReference) -> object:
         """
         Returns updated state of the object.
 
@@ -95,7 +96,7 @@ class AbstractConnectionCLIENT:
         """
         raise NotImplementedError
 
-    async def execute_method(self, object_ref: daf.misc.ObjectReference, method_name: str, **kwargs):
+    async def execute_method(self, object_ref: it.ObjectReference, method_name: str, **kwargs):
         """
         Executes a method inside object and returns the result.
 
@@ -132,8 +133,8 @@ class LocalConnectionCLIENT(AbstractConnectionCLIENT):
         if isinstance(obj, list):
             return [self._convert_ids(x) for x in obj]
 
-        if isinstance(obj, daf.misc.ObjectReference):
-            return daf.misc.get_by_id(obj.ref)
+        if isinstance(obj, it.ObjectReference):
+            return it.get_by_id(obj.ref)
 
         return obj
 
@@ -144,8 +145,8 @@ class LocalConnectionCLIENT(AbstractConnectionCLIENT):
     def add_account(self, obj: daf.client.ACCOUNT):
         return daf.add_object(obj)
 
-    def remove_account(self, account_ref: daf.misc.ObjectReference):
-        return daf.remove_object(daf.misc.get_by_id(account_ref.ref))
+    def remove_account(self, account_ref: it.ObjectReference):
+        return daf.remove_object(it.get_by_id(account_ref.ref))
 
     async def get_accounts(self) -> List[daf.client.ACCOUNT]:
         return daf.get_accounts()
@@ -153,11 +154,11 @@ class LocalConnectionCLIENT(AbstractConnectionCLIENT):
     async def get_logger(self) -> daf.logging.LoggerBASE:
         return daf.get_logger()
 
-    async def refresh(self, object_ref: daf.misc.ObjectReference):
-        return daf.misc.get_by_id(object_ref.ref)  # Local connection can just use the local object
+    async def refresh(self, object_ref: it.ObjectReference):
+        return it.get_by_id(object_ref.ref)  # Local connection can just use the local object
 
-    async def execute_method(self, object_ref: daf.misc.ObjectReference, method_name: str, **kwargs):
-        result = getattr(daf.misc.get_by_id(object_ref.ref), method_name)(**self._convert_ids(kwargs))
+    async def execute_method(self, object_ref: it.ObjectReference, method_name: str, **kwargs):
+        result = getattr(it.get_by_id(object_ref.ref), method_name)(**self._convert_ids(kwargs))
         if isinstance(result, Coroutine):
             result = await result
 
@@ -247,7 +248,7 @@ class RemoteConnectionCLIENT(AbstractConnectionCLIENT):
         )
         trace(response["message"])
 
-    async def remove_account(self, account_ref: daf.misc.ObjectReference):
+    async def remove_account(self, account_ref: it.ObjectReference):
         trace("Removing remote account.")
         response = await self._request("DELETE", "/accounts", account_id=account_ref.ref)
         trace(response["message"])
@@ -260,11 +261,11 @@ class RemoteConnectionCLIENT(AbstractConnectionCLIENT):
         response = await self._request("GET", "/logging")
         return daf.convert.convert_from_semi_dict(response["result"]["logger"])
 
-    async def refresh(self, object_ref: daf.misc.ObjectReference):
+    async def refresh(self, object_ref: it.ObjectReference):
         response = await self._request("GET", "/object", object_id=object_ref.ref)
         return daf.convert.convert_from_semi_dict(response["result"]["object"])
 
-    async def execute_method(self, object_ref: daf.misc.ObjectReference, method_name: str, **kwargs):
+    async def execute_method(self, object_ref: it.ObjectReference, method_name: str, **kwargs):
         kwargs = daf.convert.convert_object_to_semi_dict(kwargs)
         response = await self._request("POST", "/method", object_id=object_ref.ref, method_name=method_name, **kwargs)
         message = response.get("message")
